==> Script/01_shapefiles_to_xyz.Rmd <==
#Converting shapefiles to xyz via rasters

Libraries
```{r}
library(tidyverse)
library(sf)
library(raster)
library(fasterize)
library(countrycode)

select <- dplyr::select

yannick_dir <- "/rd/gem/private/users/yannickr"
fao_regions_dir <- "/rd/gem/private/users/yannickr/FAO_AREAS_CWP_NOCOASTLINE"
eez_dir <- "/rd/gem/private/users/yannickr/World_EEZ_v11_20191118"

```

Import fao regions shapefile and create numeric variable for fao areas from the original character
```{r}

fao_regions_shp <- st_read(file.path(fao_regions_dir, "FAO_AREAS_CWP_NOCOASTLINE.shp"))

fao_regions_shp <- fao_regions_shp %>%  mutate(F_area = F_AREA %>% as.numeric())


#check it plots OK using both variables (both look fine)

ggplot()+
  geom_sf(data = fao_regions_shp, aes(fill = F_AREA))


ggplot()+
  geom_sf(data = fao_regions_shp, aes(fill = F_area))

```

Some quick checks on the new variable
```{r}
unique(fao_regions_shp$F_AREA)
unique(fao_regions_shp$F_area)

# checking no new levels have been added in 

#same length?
length(unique(fao_regions_shp$F_AREA))==length(unique(fao_regions_shp$F_area))

#same values?
unique(fao_regions_shp$F_AREA)==unique(fao_regions_shp$F_area)
```

Convert shapefile to raster
```{r}

base_raster <- raster(res = 0.5)

#Use eez file as base to see if that works
eez_file <- read_csv(file.path(yannick_dir, "Cells_LatLon_EEZ.csv")) %>% select(Lon, Lat, A_Code)
eez_base_raster <- rasterFromXYZ(eez_file)
plot(eez_base_raster)


fao_raster <- fasterize(sf = fao_regions_shp, raster = eez_base_raster, field = "F_area", fun = "last")

plot(fao_raster) #looks OK - FAO area numbers are retained


#unique values (except NA) in raster should match unique areas in shapefile  
sort(unique(values(fao_raster)))==sort(unique(fao_regions_shp$F_AREA))

```

Convert raster to xyz object and export
```{r}

xyz_file <- rasterToPoints(fao_raster) %>% 
  as.data.frame() %>% 
  rename(fao_area = layer) %>% 
  mutate(fao_area = as.character(fao_area))

xyz_file <- bind_cols(raster::xyFromCell(object = fao_raster, cell = 1:ncell(fao_raster)), 
                      values(fao_raster)) %>% 
  rename(fao_area = `...3`)
  


saveRDS(object = xyz_file, file =  file.path(yannick_dir, "Cells_LatLon_FAO_regions.rds"))

```

Import EEZ shapefile
```{r}

eez_shp <- st_read(file.path(eez_dir, "eez_boundaries_v11.shp")) %>% 
  mutate(iso3n = countrycode(sourcevar = eez_shp$SOVEREIGN1, origin = "country.name", destination = "iso3n", warn = TRUE))

class(eez_shp$iso3n)

extent(eez_shp)

```

Convert shp to EEZ raster

```{r}

base_raster <- raster(res = 0.5)
extent(base_raster) <- extent(eez_shp)

eez_raster <- raster::rasterize(x = eez_shp  , y = base_raster, field = eez_shp$iso3n , fun = "last")



```
