==> Script/02_merge_effort_spatial_data_old.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Libraries

```{r}

rm(list=ls())

library(tidyverse)
library(dtplyr)
library(dplyr)
library(data.table)
library(here)
library(vroom)
library(parallel)
library(tictoc)
library(geoR)  
library(raster)
library(RColorBrewer)

select <- dplyr::select

yannick_dir <- "/rd/gem/private/users/yannickr"

original_effort_dir <- "/home/ubuntu/gem/private/users/yannickr/effort_mapped_bycountry"

aggregated_files_dir <- "/rd/gem/private/users/yannickr/effort_mapped_by_country_aggregated/"

```

# Explore files and join the main data

```{r}

(EEZ <- read_csv(file.path(yannick_dir, "Cells_LatLon_EEZ.csv"))) 

## CN how many before we filter? 
EEZ_tot<-EEZ %>% 
  filter(FAOname != "High Seas") %>% 
  select(FAOname, CNumber) %>% 
  unique()

nrow(EEZ_tot) # 193
 
#EEZ data is duplicated in places because of disputed claims to cells - just going to take the first country listed
(EEZ_disputes <- read_csv(file.path(yannick_dir, "Cells_LatLon_EEZ.csv")) %>% filter(DISPUTED==TRUE))

(take_first <- EEZ_disputes[!duplicated(EEZ_disputes[c("Lon", "Lat")]),])

head(EEZ) # CN - need to change aggregation levels - use the FAOname as the  Admin_Country aggregates EEZ that are under the same administrative country. 
unique(EEZ[,c("CNumber", "FAOname", "A_Code", "Admin_Country")])

(EEZ_adj <- bind_rows(EEZ %>% 
            filter(DISPUTED == FALSE),
          take_first) %>% 
  select(Lat, Lon, CNumber, FAOname) %>%
  rename(eez_country_code = CNumber,
         eez_country_name = FAOname) %>%
   distinct())

united_coords <- EEZ_adj %>% unite("coords", c("Lon", "Lat"), sep = " ")

EEZ_adj <- EEZ_adj[-which(duplicated(united_coords[,"coords"])==TRUE),] %>% as.data.table() # remove duplicated EEZ records
EEZ_adj <- lazy_dt(EEZ_adj)

#check EEZ data - EEZs look ok

EEZ_xyz <- EEZ_adj %>% as_tibble() %>%  rename(x=Lon, y= Lat, z=eez_country_code) %>%  select(x,y,z) 
EEZ_raster <- raster::rasterFromXYZ(EEZ_xyz)
plot(EEZ_raster)


EEZ_adj<-EEZ_adj %>% select(-eez_country_code) # CN no need for this enymore - avod confuction
nrow(as.data.table(EEZ_adj))
EEZ_adj<-unique(EEZ_adj)
nrow(as.data.table(EEZ_adj))

# ggplot(data = EEZ_xyz)+
#   geom_tile(aes(x=x, y=y, fill=z)) + 
#   scale_fill_gradientn(colours = RColorBrewer::brewer.pal(n = 9, name = "Spectral"))
# 
# ggsave(filename = here("Explore/cottrell-explore/eez_wo_duplicates_CN.jpg"), dpi=300, device = "jpg", width = 9, height = 5)

#LME data
(LMEs <- fread(file.path(yannick_dir, "LMESeq.csv")))
LMEs <- lazy_dt(LMEs)

LME_xyz <- LMEs %>% as_tibble() %>%  rename(x=Lon, y= Lat, z=LME) %>%  select(x,y,z) 
LME_raster <- raster::rasterFromXYZ(LME_xyz)
plot(LME_raster) # CN NOTE - land is 0 here.  

#FAO regions data
(FAO_regions <- readRDS(file.path(yannick_dir, "Cells_LatLon_FAO_regions.rds")) %>% 
    rename(Lon=x, Lat = y) %>% mutate(fao_area= as.double(fao_area)) %>% as.data.table())

FAO_regions <- lazy_dt(FAO_regions)

FAO_xyz <- FAO_regions %>% as_tibble() %>%  rename(x=Lon, y= Lat, z=fao_area) %>%  select(x,y,z) 
FAO_raster <- raster::rasterFromXYZ(FAO_xyz)
plot(FAO_raster) 

# for eez: Looking at the CNumber - this appears the be the iso3n code for the country. This is often the name the same as the admin country but is different when the land masses are territories of other countries

#The data have different rows. Want to make sure they line up
nrow(EEZ_adj) #179904
nrow(as_tibble(LMEs)) #259200
nrow(FAO_regions) #229369

#Some LME data are around that have no FAO or EEZ data
(all_spatial_layers <- 
  full_join(x = full_join(x = LMEs %>% select(-Seq), y = FAO_regions, by = c("Lon", "Lat")), 
            y = EEZ_adj %>% select(Lon, Lat, eez_country_name), by = c("Lon", "Lat")) %>% as_tibble()
)

missing_fao_eez <- 
  all_spatial_layers %>% 
  filter(is.na(fao_area)) %>% #29831 where FAO areas are NA but LME are not
  filter(is.na(eez_country_name)) #29830 where EEZ are NA but LME are not - this does not change depending on order. So there is one cell where the is LME data and EEZ data but not FAO area data

# Camilla's CHECKS 
# the above one grid cell needs to be fixed otherwise it is carried out in final effort dataset and giving an additional FAO region called 'NA', where LME = 0 and EEZ = GuineaBissau
# by looking at map, this is FAO 34
all_spatial_layers %>% 
  filter(is.na(fao_area) & !is.na(eez_country_name) & !is.na(LME))

a<-data.frame(Lon =-14.75, Lat=12.25, fao_area=34)
FAO_regions<-FAO_regions %>% full_join(a)
FAO_regions %>% filter(Lon == -14.75, Lat == 12.25)

# # Camilla's CHECKS
# nrow(missing_fao_eez) # 29830 - is this land, which is given as 0 for LMEs instead of NAs?
# unique(missing_fao_eez$LME) # LME 0 and LME 58  
# # check LME 0: LME 0 also includes land - see above 
# trial<-as.data.table(select(LMEs, "Lon", "Lat", "LME"))
# # check LME 58 - kara sea close to north pole.
# check_58<-filter(missing_fao_eez, LME == 58)  
# trial2<-as.data.table(filter(trial, LME == 58)) 
# nrow(check_58) # OK this is only 6 cells over 2002 belonging to kara sea
# nrow(trial2)
# # check data for LME 58
# check_58_notMissing<-filter(all_spatial_layers, LME == 58) # some Area == 0 which is wrong but we are not actually providing info on areas here. so all good.
# # FAO area and EEZ for the NAs should be: 
# unique(check_58_notMissing$fao_area) # either 18 (Arctic Sea) or 27 (Atlantic north East- adiacent to 18)
# unique(check_58_notMissing$eez_country_name) # Russian Fed
# # check map - low resolution but looking OK
# colnames(trial2)<-c("x", "y", "z")
# trial2_raster<-rasterFromXYZ(trial2)
# plot(trial2_raster)
# # decision: leave as it is ....
# 
# # why so many NAs for EEZ? 
# missing_eez <- 
#   all_spatial_layers %>% 
#   filter(is.na(eez_country_name)) 
# nrow(missing_eez) # 79296
# missing_eez2<-missing_eez %>% 
#   filter(LME != 0) # remove LME 0 and EEZ NAs which could be land and open ocean 
# nrow(missing_eez2) # 4303 - these should not be NAs: if they were outside EEZ and withing LME they should be labelled as 0s. Possible reason for discrepancy: different land-ocean mask between LME and EEZ  
# # check single EEZ/LME combination. File found in yannickr folder transformed in cvs (CN).
# FAO_LME_keys_yannick <- read_csv("/rd/gem/private/users/yannickr/LME_FAO_coding_CN.csv")
# lme_codes<-FAO_LME_keys_yannick %>% 
#   filter(LME_FAO_code == "LME") %>% 
#   select(LME_FAO_name, LME_FAO_number)
# colnames(lme_codes)<-c("LME_name", "LME")
# missing_eez2<-missing_eez2 %>% 
#   left_join(lme_codes)
# unique(missing_eez2[,c("LME_name", "LME")]) # all problematic EEZs at poles 
# # plot missing values to check: 
# missing_eez3<-missing_eez2 %>% 
#   filter(LME == 18) 
# trial<-as.data.table(select(missing_eez3, "Lon", "Lat", "LME"))
# trial_raster<-rasterFromXYZ(trial)
# plot(trial_raster)
# 
# # could be land discrepancy: 
# # 1 yes
# # 56 could be? 
# # 18 yes ... 
# 
# ## only a few LME/EEZ at high lat affected. also these are the files used to work with catch and effort by Reg and Yannick and here we are using these files for consistency. 

```

# Add interection between LME and FAO 

Mask created by Denisse but here not needed as FAO + LME column already consider this interaction  

```{r}

# # CN add LME_FAO intersections using mask provided by Denisse 
# # /rd/gem/private/fishmip_inputs/ISIMIP3a/fishmip_regions/FAO-LME_masks/
# FAO_LME <- read_csv("/rd/gem/private/fishmip_inputs/ISIMIP3a/fishmip_regions/FAO-LME_masks/FAO-LME-corrected_05degmask_DBEM.csv")
# 
# # # CHECKS 
# # head(all_spatial_layers)
# # sort(unique(all_spatial_layers$Lon)) # from -179.75 to 179.75
# # length(unique(all_spatial_layers$Lon)) # 720
# # sort(unique(all_spatial_layers$Lat)) # from -89.75 to 89.75
# # length(unique(all_spatial_layers$Lat)) # 360
# # 
# # head(FAO_LME)
# # sort(unique(FAO_LME$Lon)) # -179.75 to 179.75
# # length(unique(FAO_LME$Lon)) # 720
# # sort(unique(FAO_LME$Lat)) # from -85.25 - because Antarctica.  
# # length(unique(FAO_LME$Lat)) # 351
# # unique()
# # 
# # sort(unique(FAO_LME$ID_merged)) # See keys below 
# # 
# # #CHECK plot mask 
# # library(raster)
# # trial<-rasterFromXYZ(FAO_LME) 
# # plot(trial)
# # 
# # # check LME and FAO regions
# # sort(unique(as.data.table(LMEs)$LME)) # LME should be from 1 to 66 
# # sort(unique(as.data.table(FAO_regions)$fao_area)) # FAO area should be: 18 21 27 31 34 37 41 47 48 51 57 58 61 67 71 77 81 87 88
# # # Instead I have 83 ID because 3 FAO area overlap with LMEs (e.g. mediterranean sea - see Denisse's email on 9/05/23 RE: Catch and effort forcing for 3a)
# # length(unique(FAO_LME$ID_merged))
# 
# # codes 
# FAO_LME_keys <- read_csv("/rd/gem/private/fishmip_inputs/ISIMIP3a/fishmip_regions/FAO-LME_masks/FAO_LME_keys.csv")
# 
# # nrow(FAO_LME_keys)
# # View(FAO_LME_keys)
# # # WARNING - repetitions: 61,71,81,88
# FAO_LME_keys<-unique(FAO_LME_keys)
# # nrow(FAO_LME_keys)
# 
# # integrate info that Denisse provided on LME names 
# LME_names <- read_csv("/rd/gem/private/fishmip_inputs/ISIMIP3a/fishmip_regions/FAO-LME_masks/LME_names.csv")
# 
# # nrow(LME_names) # 66
# 
# FAO_LME_keys<-FAO_LME_keys %>% 
#   full_join(LME_names)
# 
# # combine with FAO_LME
# FAO_LME_keys_new<-FAO_LME_keys %>% 
#   mutate(Sector = ifelse(is.na(original_lme_id), "FAO", "LME"),
#          ID_original = ifelse(is.na(original_lme_id), ID_merged, original_lme_id),
#          LME_FAO = paste(Sector, ID_original, sep = "_")) %>% 
#   select(ID_merged, LME_FAO)
# 
# # nrow(FAO_LME_keys_new) # 83
# 
# FAO_LME_new<-FAO_LME %>% 
#   full_join(FAO_LME_keys_new) %>% 
#   select(-ID_merged)
# 
# # nrow(FAO_LME) # 173371
# # nrow(FAO_LME_new) # 173371
# 
# # use this file for the aggregation and check outputs
# 
# # further checks 
# # # from https://iwlearn.net/marine/lmes/list
# # 1: East Bering Sea	
# # 18: West Greenland Shelf	
# # 35: Gulf of Thailand	
# # 52: Sea of Okhotsk
# # 2: Gulf of Alaska	
# # 19: East Greenland Shelf	
# # 36: South China Sea	
# # 53: West Bering Sea
# # 3: California Current	
# # 20: Barents Sea	
# # 37: Sulu-Celebes Sea	
# # 54: Northern Bering - Chukchi Seas
# # 4: Gulf of California	
# # 21: Norwegian Sea	
# # 38: Indonesian Sea	
# # 55: Beaufort Sea
# # 5: Gulf of Mexico	
# # 22: North Sea	
# # 39: North Australian Shelf	
# # 56: East Siberian Sea
# # 6: Southeast U.S. Continental Shelf	
# # 23: Baltic Sea 	
# # 40: Northeast Australian Shelf	
# # 57: Laptev Sea
# # 7: Northeast U.S. Continental Shelf	
# # 24: Celtic-Biscay Shelf	
# # 41: East Central Australian Shelf	
# # 58: Kara Sea
# # 8: Scotian Shelf	
# # 25: Iberian Coastal	
# # 42: Southeast Australian Shelf	
# # 59: Iceland Shelf and Sea
# # 9: Newfoundland-Labrador Shelf	
# # 26: Mediterranean Sea	
# # 43: South West Australian Shelf	
# # 60: Faroe Plateau
# # 10: Insular Pacific-Hawaiian	
# # 27: Canary Current	
# # 44: West Central Austalian Shelf	
# # 61: Antarctica
# # 11: Pacific Central-American Coastal	
# # 28: Guinea Current	
# # 45: Northwest Australian Shelf	
# # 62: Black Sea
# # 12: Caribbean Sea	
# # 29: Benguela Current	
# # 46: New Zealand Shelf	
# # 63: Hudson Bay Complex
# # 13: Humboldt Current	
# # 30: Agulhas Current	
# # 47: East China Sea	
# # 64: Central Arctic
# # 14: Patagonian Shelf	
# # 31: Somali Coastal Current	
# # 48: Yellow Sea	
# # 65: Aleutian Islands
# # 15: South Brazil Shelf	
# # 32: Arabian Sea	
# # 49: Kuroshio Current	
# # 66: Canadian High Arctic - North Greenland
# # 16: East Brazil Shelf	
# # 33: Red Sea	
# # 50: Sea of Japan	
# # 17: North Brazil Shelf	
# # 34: Bay of Bengal	
# # 51: Oyashio Current	
# # 
# # LME_key_check<- data.frame(LME_number = c(1,18,35,52,2,19,36,53,3,20,37,54,4,21,38,55,5,22,39,56,6,23,40,57,7,24,41,58,8,25,42,59,9,26,43,60,10,27,44,61,11,28,45,62,12,29,46,63,13,30,47,64,14,31,48,65,15,32,49,66,16,33,50,17,34,51), name_merge = c("East Bering Sea","West Greenland Shelf","Gulf of Thailand","Sea of Okhotsk","Gulf of Alaska","East Greenland Shelf","South China Sea","West Bering Sea","California Current","Barents Sea","Sulu-Celebes Sea","Northern Bering - Chukchi Seas","Gulf of California","Norwegian Sea","Indonesian Sea","Beaufort Sea","Gulf of Mexico","North Sea","North Australian Shelf","East Siberian Sea","Southeast U.S. Continental Shelf","Baltic Sea","Northeast Australian Shelf","Laptev Sea","Northeast U.S. Continental Shelf","Celtic-Biscay Shelf","East Central Australian Shelf","Kara Sea","Scotian Shelf","Iberian Coastal","Southeast Australian Shelf","Iceland Shelf and Sea","Newfoundland-Labrador Shelf","Mediterranean Sea","South West Australian Shelf","Faroe Plateau","Insular Pacific-Hawaiian","Canary Current","West Central Austalian Shelf","Antarctica","Pacific Central-American Coastal","Guinea Current","Northwest Australian Shelf","Black Sea","Caribbean Sea","Benguela Current","New Zealand Shelf","Hudson Bay Complex","Humboldt Current","Agulhas Current","East China Sea","Central Arctic","Patagonian Shelf","Somali Coastal Current","Yellow Sea","Aleutian Islands","South Brazil Shelf", "Arabian Sea","Kuroshio Current","Canadian High Arctic - North Greenland","East Brazil Shelf", "Red Sea", "Sea of Japan","North Brazil Shelf","Bay of Bengal", "Oyashio Current"))
# # 
# # View(LME_key_check)
# # View(LME_names) # LME 18 and 19 have different names but OK 
# # 
# ### FAO AREA FROM WEBSITE - Checked all OK 
# # Area 18 # NOT IN MASK
# # ARCTIC SEA (Major Fishing Area 18)
# # Area 21
# # Northwest Atlantic (Major Fishing Area 21)
# # Area 27
# # ATLANTIC, NORTHEAST (Major Fishing Area 27)
# # Area 31
# # ATLANTIC, WESTERN-CENTRAL (Major Fishing Area 31)
# # Area 34
# # ATLANTIC, EASTERN CENTRAL (Major Fishing Area 34)
# # Area 37 # NOT IN MASK 
# # MEDITERRANEAN AND BLACK SEA (Major Fishing Area 37)
# # Area 41
# # ATLANTIC, SOUTHWEST (Major Fishing Area 41)
# # Area 47
# # ATLANTIC, SOUTHEAST (Major Fishing Area 47)
# # Area 48
# # Atlantic, Antarctic (Major Fishing Area 48)
# # Area 51
# # INDIAN OCEAN, WESTERN (Major Fishing Area 51)
# # Area 57
# # INDIAN OCEAN, EASTERN (Major Fishing Area 57)
# # Area 58
# # Antarctic and Southern Indian Ocean (Major Fishing Area 58)
# # Area 61
# # PACIFIC, NORTHWEST (Major Fishing Area 61)
# # Area 67
# # PACIFIC, NORTHEAST (Major Fishing Area 67)
# # Area 71
# # PACIFIC, WESTERN CENTRAL (Major Fishing Area 71)
# # Area 77
# # PACIFIC, EASTERN CENTRAL (Major Fishing Area 77)
# # Area 81
# # PACIFIC, SOUTHWEST (Major Fishing Area 81)
# # Area 87
# # PACIFIC, SOUTHEAST (Major Fishing Area 87)
# # Area 88
# # Pacific, Antarctic (Major Fishing Area 88)

```

# Add interactions between EEZ and FAO 

Not needed see reason above 

```{r}

# FAO_EEZ_keys <- read_csv("/rd/gem/private/fishmip_inputs/ISIMIP3a/fishmip_regions/FAO_EEZ_masks/FAO_EEZ_keys.csv")
# 
# FAO_EEZ_keys<-FAO_EEZ_keys %>% mutate(name_merge = gsub(" Exclusive Economic Zone", "", name_merge), 
#                         name_merge = gsub(" Exclusive economic Zone", "", name_merge),
#                         name_merge = gsub(" Exclusive Economi", "", name_merge),
#                         name_merge = gsub(" Exclusive Econo", "", name_merge))
# 
# 
# # check only EEZ 
# trial<-FAO_EEZ_keys[19:nrow(FAO_EEZ_keys),]
# trial<-unique(trial)
# nrow(trial) # 281 EEZ 
# sort(unique(trial$name_merge))
# 
# # check with EEZ from Yannick and used for merging 
# trial2<-EEZ_adj %>% filter(eez_country_code != 0) %>% select(eez_country_code, eez_country_name) %>% unique()
# 
# nrow(as.data.table(trial2)) # 186
# sort(unique(as.data.table(trial2)$eez_country_name))
# 
# # SEND EMAIL TO DENISSE - DONE  

```

# Run function to aggregate data on all 30,000 files and write them to folder

```{r}

source("Script/Helper_functions_fishingEffort.R")

#compressed_effort_files <- list.files(file.path(compressed_effort_dir), full.names = TRUE)
original_effort_files <- list.files(file.path(original_effort_dir), pattern = ".csv")

chunk_size <- 500 #chunk size for processing
effort_list_split <- split(original_effort_files, ceiling(seq_along(original_effort_files)/chunk_size))
length(effort_list_split) #60 chunks

tic()
for(i in 1:length(effort_list_split)){
  
  file_chunk <- effort_list_split[[i]]
  
  message("Processing chunk #", i, " of ", length(effort_list_split))
  
  mclapply(X = file_chunk, FUN = join_effort_data, mc.cores = 40)
  
}
toc()
4679.399/3600 #1.3 hours to run

```

# Check the files

```{r}


#check the files
# newly_written_files <- list.files(file.path(yannick_dir, "effort_mapped_by_country_aggregated"), full.names = TRUE)

# # CN consider the EEZ version as explained above 
newly_written_files <- list.files(file.path(yannick_dir, "effort_mapped_by_country_aggregated_EEZ"), full.names = TRUE)

#pick one randomly 
map(newly_written_files[[8]], fread)
```

# Combine all listed files to one file

```{r}

combined_aggregated_effort <- rbindlist(mclapply(X = newly_written_files, FUN = fread, mc.cores = 40))

# fwrite(x = combined_aggregated_effort, file.path(yannick_dir, "all_effort_aggregated.csv"))

# CN consider the EEZ version as explained above
fwrite(x = combined_aggregated_effort, file.path(yannick_dir, "all_effort_aggregated_EEZ.csv"))

```
